+++ client/package.json
@@ -12,7 +12,7 @@
     "jwt-decode": "^2.2.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-router-dom": "^5.1.2",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "^3.4.1",
     "web-vitals": "^2.1.4"
   },
