+++ client/src/App.js
@@ -1,23 +1,87 @@
+import React from 'react'
+import {
+  ApolloClient,
+  InMemoryCache,
+  ApolloProvider,
+  createHttpLink,
+} from "@apollo/client";
+import { setContext } from "@apollo/client/link/context";
+// import FlashcardList from './LanguageCards/FlashcardList';
+// import {spanishList} from './LanguageCards/spanishDB';
+// import {frenchList} from './LanguageCards/frenchDB';
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import NavbarOutline from "./components/Navbar/Navbar";
+import Footer from "./components/Footer";
+// import PlanTrip from "./PlanTrip";
 
-import './App.css';
-import React, {useState} from 'react';
-import FlashcardList from './LanguageCards/FlashcardList';
-import {spanishList} from './LanguageCards/spanishDB';
-import {frenchList} from './LanguageCards/frenchDB';
-import {BrowserRouter as Router, route } from "react-router-dom";
+import Home from "./pages/Home";
+import Signup from "./pages/Signup";
+import Login from "./pages/Login";
+import Profile from './pages/Profile';
 
-import NavbarOutline from './NavbarOutline';
+// Construct our main GraphQL API endpoint
+const httpLink = createHttpLink({
+  uri: "/graphql",
+});
+
+// Construct request middleware that will attach the JWT token to every request as an `authorization` header
+const authLink = setContext((_, { headers }) => {
+  // get the authentication token from local storage if it exists
+  const token = localStorage.getItem("id_token");
+  // return the headers to the context so httpLink can read them
+  return {
+    headers: {
+      ...headers,
+      authorization: token ? `Bearer ${token}` : "",
+    },
+  };
+});
+
+const client = new ApolloClient({
+  // Set up our client to execute the `authLink` middleware prior to making the request to our GraphQL API
+  link: authLink.concat(httpLink),
+  cache: new InMemoryCache(),
+});
+
 function App() {
-  const [cards, setCards] = useState(frenchList)
   return (
+    <ApolloProvider client={client}>
     <Router>
-    <div>
-      
-     <NavbarOutline/>
-
-      <FlashcardList flashcards = {cards}/>
-    </div>
+    <div className="flex-column justify-flex-start min-100-vh">
+        <NavbarOutline />
+        {/* <FlashcardList flashcards = {cards}/> */}
+        <div className="container">
+          <Routes>
+            {/* <Route 
+            path="/home" 
+            component={PlanTrip} 
+            /> */}
+              <Route 
+                path="/"
+                element={<Home />}
+              />
+              <Route 
+                path="/login" 
+                element={<Login />}
+              />
+              <Route 
+                path="/signup" 
+                element={<Signup />}
+              />
+              <Route 
+                path="/me" 
+                element={<Profile />}
+              />
+              <Route 
+                path="/profiles/:username" 
+                element={<Profile />}
+              />
+              </Routes>
+          </div>
+          <Footer />
+        </div>
     </Router>
+    </ApolloProvider>
   );
 }
