+++ client/src/pages/Home.js
@@ -1,8 +1,7 @@
 import React from 'react';
-import { useQuery } from '@apollo/client';
+// import { useQuery } from '@apollo/client';
 
 const Home = () => {
-  const { loading, data } = useQuery();
 
   return (
     <main>
