+++ package.json
@@ -14,6 +14,8 @@
     "author": "",
     "license": "ISC",
     "dependencies": {
-        "concurrently": "^7.6.0"
+        "@apollo/client": "^3.7.1",
+        "concurrently": "^7.6.0",
+        "react-router-dom": "^6.2.1"
     }
 }
